# btf: use iterators